04_DiversTests/03_Application_Labelbox/POC_Imptox_tab/requirements_for_POC.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
affine=2.3.0=py_0
aiohttp=3.8.1=py38h294d835_0
aiosignal=1.2.0=pyhd8ed1ab_0
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backoff=1.10.0=pypi_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=22.1.0=pyhd8ed1ab_0
blosc=1.21.0=h0e60522_0
boost-cpp=1.74.0=h5b4e17d_5
branca=0.4.2=pyhd8ed1ab_0
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py38h294d835_1003
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cachetools=5.0.0=pypi_0
cairo=1.16.0=h15b3021_1009
certifi=2021.10.8=py38haa244fe_1
cffi=1.15.0=py38hd8c33c5_0
cfitsio=4.0.0=hd67004f_0
charset-normalizer=2.0.11=pyhd8ed1ab_0
click=8.0.3=py38haa244fe_1
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38h21b164f_0
curl=7.81.0=h789b8ee_0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.5.1=py38h885f38d_0
decorator=4.4.2=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
executing=0.8.2=pyhd8ed1ab_0
expat=2.4.4=h39d44d4_0
fiona=1.8.20=py38h19672d5_4
flatbuffers=2.0=pypi_0
folium=0.12.1.post1=pyhd8ed1ab_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.94=h1989441_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.29.1=py38h294d835_0
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
frozenlist=1.3.0=py38h294d835_0
gast=0.4.0=pypi_0
gdal=3.4.0=py38h67fab55_12
geojson=2.5.0=pypi_0
geopandas=0.10.2=pyhd8ed1ab_1
geopandas-base=0.10.2=pyha770c72_1
geos=3.10.1=h39d44d4_1
geotiff=1.7.0=hc8731e1_5
gettext=0.19.8.1=ha2e2712_1008
google-api-core=2.4.0=pypi_0
google-auth=2.6.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pypi_0
google-cloud=0.34.0=pypi_0
google-cloud-core=2.2.2=pypi_0
google-cloud-storage=2.1.0=pypi_0
google-crc32c=1.3.0=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=2.1.0=pypi_0
googleapis-common-protos=1.54.0=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
hdf4=4.2.15=h0e5069d_3
hdf5=1.12.1=nompi_h57737ce_103
icu=69.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
imagecodecs=2020.5.30=pypi_0
imageio=2.9.0=pypi_0
imagesize=1.3.0=pypi_0
importlib-metadata=4.10.1=pypi_0
intel-openmp=2022.0.0=h57928b3_3663
ipykernel=6.7.0=py38h595d716_0
ipython=8.0.1=py38haa244fe_0
jbig=2.1=h8d14728_2003
jcamp=1.2.1=pypi_0
jedi=0.18.1=py38haa244fe_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.1=py38haa244fe_1
kealib=1.4.14=h8995ca9_3
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=pypi_0
krb5=1.19.2=h6da9e4a_3
labelbox=3.12.0=pypi_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=13_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libcblas=3.9.0=13_win64_mkl
libclang=13.0.0=pypi_0
libcurl=7.81.0=h789b8ee_0
libdeflate=1.8=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libgdal=3.4.0=h58f6a35_12
libglib=2.70.2=h3be07f2_1
libiconv=1.16=he774522_0
libkml=1.3.0=h9859afa_1014
liblapack=3.9.0=13_win64_mkl
libnetcdf=4.8.1=nompi_h1cc8e9d_101
libpng=1.6.37=h1d00b33_2
libpq=14.1=h1ea2d34_1
libprotobuf=3.19.4=h7755175_0
librttopo=1.1.0=he35e8ac_8
libsodium=1.0.18=h8d14728_1
libspatialindex=1.9.3=h39d44d4_4
libspatialite=5.0.1=hf126459_12
libssh2=1.10.0=h9a1e1f7_2
libtiff=4.3.0=hd413186_2
libwebp-base=1.2.2=h8ffe710_1
libxml2=2.9.12=hf5bbc77_1
libzip=1.8.0=h519de47_1
libzlib=1.2.11=h8ffe710_1013
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mapclassify=2.4.3=pyhd8ed1ab_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py38h294d835_1
matplotlib=3.3.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2022.0.0=h0e2418a_796
msys2-conda-epoch=20160418=1
multidict=6.0.2=py38h294d835_0
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
mypy_extensions=0.4.3=py38haa244fe_4
mysql-connector-python=8.0.21=pypi_0
ndjson=0.3.1=pypi_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
networkx=2.5=pypi_0
numpy=1.19.1=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.4.0.42=pypi_0
openjpeg=2.4.0=hb211442_1
openssl=3.0.0=h8ffe710_2
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.1.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.9.0=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=7.2.0=pypi_0
pip=22.0.2=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
platformdirs=2.3.0=pyhd8ed1ab_0
poppler=21.11.0=h24fffdf_0
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.1=he353ca9_1
proj=8.2.0=h1cfcee9_0
prompt-toolkit=3.0.26=pyha770c72_0
protobuf=3.13.0=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.9.0=pypi_0
pygeotile=1.0.6=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
pyproj=3.3.0=py38h12f9802_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pysocks=1.7.1=py38haa244fe_4
python=3.8.12=h900ac77_2_cpython
python-dateutil=2.8.1=pypi_0
python_abi=3.8=2_cp38
pytz=2020.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
pywin32=303=py38h294d835_0
pyyaml=5.3.1=py38he774522_1
pyzmq=22.3.0=py38h09162b1_1
rasterio=1.2.10=py38h48edd3a_3
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pyhd8ed1ab_0
rtree=0.9.7=py38h8b54edf_3
scikit-fuzzy=0.4.2=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=1.0.2=py38hb60ee80_0
scipy=1.5.2=pypi_0
setuptools=49.6.0=pypi_0
shapely=1.8.0=py38hd784ac2_4
six=1.15.0=pypi_0
snuggs=1.4.7=py_0
sqlite=3.37.0=h8ffe710_0
stack_data=0.1.4=pyhd8ed1ab_0
tbb=2021.5.0=h2d74725_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.7.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2020.8.13=pypi_0
tiledb=2.5.3=h47404fa_0
tk=8.6.11=h8ffe710_1
tomli=2.0.0=pyhd8ed1ab_1
tornado=6.1=py38h294d835_2
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typed-ast=1.5.2=py38h294d835_0
typeguard=2.13.3=pypi_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
unicodedata2=14.0.0=py38h294d835_0
urllib3=1.26.8=pyhd8ed1ab_1
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py38haa244fe_3
wrapt=1.13.3=pypi_0
xerces-c=3.2.3=h0e60522_4
xlsxwriter=1.3.3=pypi_0
xyzservices=2022.1.1=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yarl=1.7.2=py38h294d835_1
zeromq=4.3.4=h0e60522_1
zipp=3.7.0=pypi_0
zlib=1.2.11=h8ffe710_1013
zstd=1.5.2=h6255e5f_0
